src/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
backcall==0.2.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.6
django-heroku==0.3.1
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
ipython==7.20.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
Markdown==3.3.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
packaging==20.9
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
Pygments==2.8.0
pyparsing==2.4.7
pytz==2021.1
regex==2020.11.13
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
whitenoise==5.2.0
